Invalid Template - Fixed
Hi, I was using this template while implementing this - https://aws.amazon.com/blogs/mt/automate-account-creation-and-resource-provisioning-using-aws-service-catalog-aws-organizations-and-aws-lambda/

I found an error in this template that was causing the Cloudformation stack creation to fail, after some debugging I figured it out. Basically, the default value in the parameter has to be one of the allowed values, or CloudFormation deems the template invalid.